requirements.txt
Flask
gunicorn
Jinja2
Werkzeug
numpy
pandas
scipy
scikit-learn
matplotlib
mysql-connector-python